VACMS-10262 refactor wysiwyg and rich text char limit 1000
