+++ src/data/delegators/entityMetaProvider.tsx
@@ -1,6 +1,4 @@
 import { DrupalJsonApiParams } from 'drupal-jsonapi-params'
-import { Meta as WysiwygMeta } from '@/templates/components/wysiwyg/dataService'
-import { Meta as RichTextCharLimit1000Meta } from '@/templates/components/richTextCharLimit1000/dataService'
 
 export interface EntityMetaInfo {
   /** Identifier for a Drupal data object. These are of the form `entity_type--entity_bundle`, for example `node--news_story` or `paragraph--email_contact`. */
@@ -35,8 +33,6 @@ export interface EntityMetaOut {
 const EntityMetaIn: EntityMetaInfo[] = [
   // blocks
   // paragraphs
-  WysiwygMeta,
-  RichTextCharLimit1000Meta,
 ]
 
 /** Converts the meta information into a form indexed by resource type. Very possibly overwrought. */
