--- src/templates/components/richTextCharLimit1000/index.tsx
@@ -1,15 +1,0 @@
-import { Wysiwyg } from '@/templates/components/wysiwyg'
-
-export interface RichTextCharLimit1000Props {
-  id: string
-  html: any
-  className?: string
-}
-
-export const RichTextCharLimit1000 = ({
-  id,
-  html,
-  className,
-}: RichTextCharLimit1000Props): JSX.Element => {
-  return <Wysiwyg id={id} html={html} className={className} />
-}
